# Align max op nan propagation with NumPy