--- a/tests/ops_tests.cpp
+++ b/tests/ops_tests.cpp
@@ -1024,6 +1024,10 @@ TEST_CASE("test reduction ops") {
     x = array({true, true, true, false, true, false}, {2, 3});
     CHECK(array_equal(min(x, 1), array({true, false})).item<bool>());
     CHECK(array_equal(min(x, 0), array({false, true, false})).item<bool>());
+
+    x = array({1.0f, NAN, 3.0f});
+    CHECK(isnan(max(x).item<float>()));
+
   }
 
   // Test logsumexp