@jhavukainen 2025-07-09
@awni, @angeloskath  Thanks for the initial review and comments! I implemented the fix in cleaning up the namespace using the alias and the check for non-integral types.

I gave the CPU implementation the same treatment and tested that it passes the new nan propagation tests for max when running with `DEVICE=cpu` env variable. While testing I found a small fib in the int8 and int16 dispatch dtypes for max on the CPU that made the max tests fail if the input array contained both negative and positive values, so I added a fix there.