@jhavukainen 2025-07-09
If all looks good, this can be merged in and I'll prepare another one for the min op.