@awni 2025-07-09
> While testing I found a small fib in the int8 and int16 dispatch dtypes for max on the CPU that made the max tests fail if the input array contained both negative and positive values, so I added a fix ther

Good catch!