@jhavukainen 2025-07-09
Changed the tuple unpacking to use more friendly syntax for earlier python versions, that should fix the syntax issue.

I think I'll need to brush up on cuda a bit before implementing it on that side as well. Added the tests to the cuda skiplist for now.